# app/Models/Driver.php
<?php

namespace App\Models;

use Illuminate\Database\Eloquent\Model;

class Driver extends Model
{
    protected $fillable = ['name'];

    public function cars()
    {
        return $this->hasMany(Car::class);
    }
}